+++ lib/app/models/user_settings.dart
@@ -47,6 +47,7 @@ class DefaultUserSettings {
   static const scrollThumbHeight = 48.0;
   static const scrollThumbWidth = 26.0;
   static const showMiniPlayerSeekButtons = false;
+  static const PlayerBackgroundTheme playerBackgroundTheme = .none;
 }
 
 @freezed
@@ -141,6 +142,9 @@ sealed class UserSettings with _$UserSettings {
 
     @Default(DefaultUserSettings.showMiniPlayerSeekButtons)
     bool showMiniPlayerSeekButtons,
+
+    @Default(DefaultUserSettings.playerBackgroundTheme)
+    PlayerBackgroundTheme playerBackgroundTheme,
   }) = _UserSettings;
 
   factory UserSettings.fromJson(Map<String, dynamic> json) =>