+++ lib/app/models/user_settings.g.dart
@@ -117,6 +117,12 @@ _UserSettings _$UserSettingsFromJson(
   showMiniPlayerSeekButtons:
       json['showMiniPlayerSeekButtons'] as bool? ??
       DefaultUserSettings.showMiniPlayerSeekButtons,
+  playerBackgroundTheme:
+      $enumDecodeNullable(
+        _$PlayerBackgroundThemeEnumMap,
+        json['playerBackgroundTheme'],
+      ) ??
+      DefaultUserSettings.playerBackgroundTheme,
 );
 
 Map<String, dynamic> _$UserSettingsToJson(
@@ -163,6 +169,8 @@ Map<String, dynamic> _$UserSettingsToJson(
   'scrollThumbHeight': instance.scrollThumbHeight,
   'scrollThumbWidth': instance.scrollThumbWidth,
   'showMiniPlayerSeekButtons': instance.showMiniPlayerSeekButtons,
+  'playerBackgroundTheme':
+      _$PlayerBackgroundThemeEnumMap[instance.playerBackgroundTheme]!,
 };
 
 const _$NavTargetEnumMap = {
@@ -241,4 +249,11 @@ const _$FullPlayerActionsEnumMap = {
   FullPlayerActions.speed: 'speed',
   FullPlayerActions.volume: 'volume',
   FullPlayerActions.bookmarks: 'bookmarks',
+};
+
+const _$PlayerBackgroundThemeEnumMap = {
+  PlayerBackgroundTheme.none: 'none',
+  PlayerBackgroundTheme.colored: 'colored',
+  PlayerBackgroundTheme.gradient: 'gradient',
+  PlayerBackgroundTheme.blur: 'blur',
 };