+++ lib/features/player/logic/playback_history.dart
@@ -5,7 +5,7 @@ import 'package:storii/storage/local/playback_history_store.dart';
 
 part 'playback_history.g.dart';
 
-@riverpod
+@Riverpod(keepAlive: true)
 class PlaybackHistoryNotifier extends _$PlaybackHistoryNotifier {
   PlaybackHistoryStore get _store =>
       ref.read(playbackHistoryStoreProvider.notifier);