+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - setting to show/hide seek buttons in miniplayer
+- player background theme options
 
 ### Fixed